@@ -13,7 +13,7 @@
   },
   "main": "dist/cjs.js",
   "engines": {
-    "node": ">= 12.13.0"
+    "node": ">= 18.12.0"
   },
   "scripts": {
     "start": "npm run build -- -w",
@@ -54,52 +61,53 @@
     "sass-embedded": {
       "optional": true
     },
-    "fibers": {
+    "webpack": {
       "optional": true
     }
   },
   "dependencies": {
-    "klona": "^2.0.4",
     "neo-async": "^2.6.2"
   },
   "devDependencies": {
-    "@babel/cli": "^7.17.0",
-    "@babel/core": "^7.17.0",
-    "@babel/preset-env": "^7.16.11",
-    "@commitlint/cli": "^16.2.1",
-    "@commitlint/config-conventional": "^16.2.1",
+    "@babel/cli": "^7.23.4",
+    "@babel/core": "^7.24.0",
+    "@babel/preset-env": "^7.24.0",
+    "@commitlint/cli": "^19.3.0",
+    "@commitlint/config-conventional": "^19.2.2",
     "@webpack-contrib/eslint-config-webpack": "^3.0.0",
-    "babel-jest": "^27.5.0",
+    "babel-jest": "^29.6.2",
     "bootstrap-sass": "^3.4.1",
     "bootstrap-v4": "npm:bootstrap@^4.5.3",
     "bootstrap-v5": "npm:bootstrap@^5.0.1",
     "cross-env": "^7.0.3",
-    "css-loader": "^6.6.0",
-    "del": "^6.0.0",
-    "del-cli": "^4.0.1",
-    "enhanced-resolve": "^5.8.2",
-    "eslint": "^8.8.0",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-import": "^2.25.4",
-    "fibers": "^5.0.1",
+    "cspell": "^8.6.0",
+    "css-loader": "^6.9.0",
+    "del": "^6.1.1",
+    "del-cli": "^5.1.0",
+    "enhanced-resolve": "^5.15.1",
+    "eslint": "^8.57.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-import": "^2.28.0",
     "file-loader": "^6.2.0",
-    "foundation-sites": "^6.6.3",
-    "husky": "^7.0.1",
-    "jest": "^27.5.0",
-    "jest-environment-node-single-context": "^27.2.0",
-    "lint-staged": "^12.3.3",
-    "material-components-web": "^8.0.0",
-    "memfs": "^3.4.1",
-    "node-sass": "^7.0.1",
+    "foundation-sites": "^6.7.5",
+    "husky": "^9.1.3",
+    "jest": "^29.6.2",
+    "jest-environment-node-single-context": "^29.1.0",
+    "lint-staged": "^15.2.0",
+    "material-components-web": "^9.0.0",
+    "memfs": "^4.7.7",
+    "node-sass": "^9.0.0",
     "node-sass-glob-importer": "^5.3.2",
     "npm-run-all": "^4.1.5",
-    "prettier": "^2.3.2",
-    "sass": "^1.49.7",
-    "sass-embedded": "^1.49.7",
-    "semver": "^7.3.5",
+    "prettier": "^3.2.2",
+    "sass": "^1.71.1",
+    "sass-embedded": "^1.71.1",
+    "semver": "^7.5.4",
     "standard-version": "^9.3.1",
-    "style-loader": "^3.2.1",
-    "webpack": "^5.68.0"
+    "style-loader": "^3.3.4",
+    "webpack": "^5.93.0",
+    "webpack-cli": "^5.1.4",
+    "webpack-dev-server": "^5.0.4"
   },
   "keywords": [
     "sass",
