@@ -7,7 +7,7 @@ interface MessageProps {
 }
 
 const Message: React.FC<MessageProps> = ({ message }) => {
-  const isOwnMessage = message.sender === 'You'; // Adjust based on actual user
+  const isOwnMessage = message.sender === 'You'; 
 
   return (
     <div className={isOwnMessage ? styles.ownMessage : styles.otherMessage}>
